requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-seed==0.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
